[graph] Split Nodes.cpp into Node/Nodes following headers
